+++ src/events/events.ts
@@ -1,8 +1,8 @@
-import { basicEvents } from "./basic";
-import { characterEvents } from "./characters";
-import { dollEvents } from "./dolls";
+import { BasicEventParameters, basicEvents } from "./basic";
+import { characterEvents, CharacterEventParameters } from "./characters";
+import { DollEventParameters, dollEvents } from "./dolls";
 import { playerEvents } from "./players";
-import { sceneEvents } from "./scene";
+import { SceneEventParameters, sceneEvents } from "./scene";
 import { soundEvents } from "./sound";
 import { speechEvents } from "./speech";
 import { stickerEvents } from "./stickers";
@@ -17,3 +17,9 @@ export const prendyEventGroups = {
   speechEvents,
   stickerEvents,
 };
+
+export type PrendyEventParameters<T_Group, T_Event, T_GenericParamA> =
+  | BasicEventParameters<T_Group, T_Event, T_GenericParamA>
+  | CharacterEventParameters<T_Group, T_Event, T_GenericParamA>
+  | DollEventParameters<T_Group, T_Event, T_GenericParamA>
+  | SceneEventParameters<T_Group, T_Event, T_GenericParamA>;