+++ src/events/speech.ts
@@ -1,36 +1,19 @@
+import { CSSProperties } from "react";
+import { getState, onNextTick, setState, startNewItemEffect, stopNewEffect } from "repond";
+import { II, chainDo, makeEventTypes, runEvents, setLiveEventState } from "repond-events";
+import { length } from "stringz";
 import { getCharDollStuff } from "../helpers/prendyUtils/characters";
 import { getGlobalState, setGlobalState } from "../helpers/prendyUtils/global";
 import { getTypingDelayForText } from "../helpers/prendyUtils/speechBubbles";
 import { meta } from "../meta";
 import { CharacterName, DollName, SpeechBubbleName } from "../types";
-import { CSSProperties } from "react";
-import { getState, onNextTick, setState, startNewItemEffect, stopNewEffect } from "repond";
-import { chainDo, makeEventTypes, runEvents, setLiveEventState, toDo } from "repond-events";
-import { length } from "stringz";
-
-// TODO update the timeouts to use repond's new timeout system
-
-type ATimeout = ReturnType<typeof setTimeout> | undefined;
-
-type CharacterNameUntyped = string;
-
-type ShowSpeechOptions = {
-  time?: number;
-  showOnce?: boolean;
-  character?: SpeechBubbleName & CharacterName; // NOTE SpeechBubble names and CharacterNames match at the moment
-  zoomAmount?: number;
-  lookAtPlayer?: boolean;
-  returnToZoomBeforeConversation?: boolean; // remembers the previous zoom instead of going to the default when the convo ends
-  // TODO rename stylesByKeyword or keywordStyles
-  stylesBySpecialText?: Record<string, CSSProperties>; // { "golden banana": { color: "yellow" } } // style snippets of text
-};
 
 const RESET_CAMERA_FOCUS_CHAIN_ID = "resetCameraFocus";
 const HIDE_ALARM_TEXT_CHAIN_ID = "hideAlarmText";
 const getCloseCharacterSpeechBubbleChainId = (character: CharacterName) => `closeSpeechBubble_${character}`;
 const getCloseCharacterMiniBubbleChainId = (character: CharacterName) => `closeMiniBubble_${character}`;
 
-const showSpeechRefs = {
+const speechRefs = {
   shownTextBools: {} as Record<string, boolean>, // { ["hello"] : true }
   aSpeechIsShowing: false, // NOTE probably better as global state or refs
   originalZoomAmount: 1,
@@ -42,76 +25,64 @@ const SPEECH_CLOSE_DELAY = 700; // close if no more messages from the character 
 const MIN_AUTO_SPEECH_TIME = 1500;
 
 export const speechEvents = makeEventTypes(({ event }) => ({
-  showSpeech: event({
-    run: async ({ text, options }, { runMode, liveId, isFirstAdd, elapsedTime }) => {
+  say: event({
+    run: async ({ what: text, ...options }, { runMode, liveId, isFirstAdd, elapsedTime }) => {
       const playerCharacter = getGlobalState().playerCharacter as CharacterName;
       const {
         time, // time = 2600,
         showOnce = false,
-        character = playerCharacter,
+        who = playerCharacter,
         zoomAmount = SPEECH_ZOOM_AMOUNT,
         returnToZoomBeforeConversation: returnToZoomBeforeConversation = false,
         stylesBySpecialText,
       } = options ?? {};
 
-      const closeCharacterSpeechChainId = getCloseCharacterSpeechBubbleChainId(character);
-
-      // const effectId = "showSpeech_handlePressButton" + Math.random();
-      const progressButtonEffectId = "showSpeech_handlePressButton" + liveId;
+      const closeCharacterSpeechChainId = getCloseCharacterSpeechBubbleChainId(who);
+      const progressButtonEffectId = "say_handlePressButton" + liveId;
 
       if (runMode === "end") {
         stopNewEffect(progressButtonEffectId);
         chainDo("cancel", closeCharacterSpeechChainId);
         chainDo("cancel", RESET_CAMERA_FOCUS_CHAIN_ID);
 
-        runEvents(
-          [
-            toDo("basic", "wait", { duration: SPEECH_CLOSE_DELAY }),
-            toDo("speech", "hideSpeech", { character, returnToZoomBeforeConversation }),
-          ],
-          { chainId: closeCharacterSpeechChainId }
-        );
+        runEvents([II("basic", "wait", { time: SPEECH_CLOSE_DELAY / 1000 }), II("speech", "hideSay", { who })], {
+          chainId: closeCharacterSpeechChainId,
+        });
 
         runEvents(
-          [toDo("basic", "wait", { duration: SPEECH_CLOSE_DELAY }), toDo("speech", "resetCameraFocus", { character })],
-          {
-            chainId: RESET_CAMERA_FOCUS_CHAIN_ID,
-          }
+          [
+            II("basic", "wait", { time: SPEECH_CLOSE_DELAY / 1000 }),
+            II("speech", "resetCameraFocus", { who, returnToZoomBeforeConversation }),
+          ],
+          { chainId: RESET_CAMERA_FOCUS_CHAIN_ID }
         );
       }
 
       if (runMode !== "start") return;
-
       const endEvent = () => setState({ liveEvents: { [liveId]: { goalEndTime: 0 } } });
 
-      if (showOnce && showSpeechRefs.shownTextBools[text]) endEvent();
+      if (showOnce && speechRefs.shownTextBools[text]) endEvent();
 
       const { prendyOptions } = meta.assets!;
-
       const { slateZoom: initialSlateZoom } = getGlobalState();
+      const { dollName } = getCharDollStuff(who);
 
-      const { dollName } = getCharDollStuff(character);
-      const { dollName: playerDollName } = getCharDollStuff(playerCharacter);
+      const timeInMilliseconds = time ? time * 1000 : undefined;
 
       // NOTE at the moment CharacterName and SpeechBubbleName are the same
-      const timeBasedOnText = MIN_AUTO_SPEECH_TIME + getTypingDelayForText(text, character as any) * 2;
-      const waitTime = time ?? timeBasedOnText;
+      const timeBasedOnText = MIN_AUTO_SPEECH_TIME + getTypingDelayForText(text, who as any) * 2;
+      const waitTime = timeInMilliseconds ?? timeBasedOnText;
 
-      // setLiveEventState(liveId, { goalEndTime: elapsedTime + waitTime });
-      setState({ liveEvents: { [liveId]: { goalEndTime: elapsedTime + waitTime } } });
-
-      // TODO set the end time based on the elpased time and the waitTime
+      setLiveEventState(liveId, { goalEndTime: elapsedTime + waitTime });
 
       function handlePressProgressButton() {
-        console.log("handlePressProgressButton");
-
-        const speechBubbleState = getState().speechBubbles[character];
+        const speechBubbleState = getState().speechBubbles[who];
         const { typingFinished, goalText } = speechBubbleState;
 
-        // typing done!
         if (!typingFinished) {
-          setState({ speechBubbles: { [character]: { visibleLetterAmount: length(goalText) } } });
-          // reading done!
+          // Finish typing
+          setState({ speechBubbles: { [who]: { visibleLetterAmount: length(goalText) } } });
+          // Finished reading
         } else endEvent();
       }
 
@@ -127,16 +98,16 @@ export const speechEvents = makeEventTypes(({ event }) => ({
       chainDo("cancel", closeCharacterSpeechChainId);
       chainDo("cancel", RESET_CAMERA_FOCUS_CHAIN_ID);
 
-      if (!showSpeechRefs.aSpeechIsShowing) {
-        showSpeechRefs.originalZoomAmount = initialSlateZoom;
-        showSpeechRefs.aSpeechIsShowing = true;
+      if (!speechRefs.aSpeechIsShowing) {
+        speechRefs.originalZoomAmount = initialSlateZoom;
+        speechRefs.aSpeechIsShowing = true;
       }
 
-      const newSlateZoom = Math.min(showSpeechRefs.originalZoomAmount * zoomAmount, prendyOptions.zoomLevels.max);
+      const newSlateZoom = Math.min(speechRefs.originalZoomAmount * zoomAmount, prendyOptions.zoomLevels.max);
       onNextTick(() => {
         setState({
           speechBubbles: {
-            [character]: {
+            [who]: {
               isVisible: true,
               goalText: text,
               stylesBySpecialText,
@@ -147,30 +118,40 @@ export const speechEvents = makeEventTypes(({ event }) => ({
           global: { main: { focusedDoll: dollName, slateZoomGoal: newSlateZoom + Math.random() * 0.001 } },
         });
       });
-      showSpeechRefs.shownTextBools[text] = true;
+      speechRefs.shownTextBools[text] = true;
     },
-    params: { text: "", options: undefined as undefined | ShowSpeechOptions },
+    params: {
+      what: "",
+      time: undefined as number | undefined,
+      showOnce: undefined as boolean | undefined,
+      who: undefined as undefined | (SpeechBubbleName & CharacterName), // NOTE SpeechBubble names and CharacterNames match at the moment
+      zoomAmount: undefined as number | undefined,
+      lookAtPlayer: undefined as boolean | undefined,
+      returnToZoomBeforeConversation: undefined as boolean | undefined, // remembers the previous zoom instead of going to the default when the convo ends
+      // TODO rename stylesByKeyword or keywordStyles
+      stylesBySpecialText: undefined as Record<string, CSSProperties> | undefined, // { "golden banana": { color: "yellow" } } // style snippets of text
+    },
     duration: Infinity,
   }),
-  hideSpeech: event({
-    run: ({ character, returnToZoomBeforeConversation }, { runMode }) => {
+  hideSay: event({
+    run: ({ who }, { runMode }) => {
       if (runMode !== "start") return;
-      setState({ speechBubbles: { [character]: { isVisible: false } } });
+      setState({ speechBubbles: { [who]: { isVisible: false } } });
     },
-    params: { character: "" as CharacterName, returnToZoomBeforeConversation: false },
+    params: { who: "" as CharacterName },
   }),
   resetCameraFocus: event({
-    run: ({ character, returnToZoomBeforeConversation }, { runMode }) => {
+    run: ({ who, returnToZoomBeforeConversation = false }, { runMode }) => {
       if (runMode !== "start") return;
 
       const { prendyOptions } = meta.assets!;
 
       const playerCharacter = getGlobalState().playerCharacter as CharacterName;
 
-      const { dollName } = getCharDollStuff(character);
+      const { dollName } = getCharDollStuff(who);
       const { dollName: playerDollName } = getCharDollStuff(playerCharacter);
 
-      showSpeechRefs.aSpeechIsShowing = false;
+      speechRefs.aSpeechIsShowing = false;
 
       const currentFocusedDoll = getGlobalState().focusedDoll;
       const isFocusedOnTalkingCharacter = currentFocusedDoll === dollName;
@@ -179,51 +160,53 @@ export const speechEvents = makeEventTypes(({ event }) => ({
       setGlobalState({
         focusedDoll: isFocusedOnTalkingCharacter ? playerDollName : currentFocusedDoll,
         slateZoomGoal: returnToZoomBeforeConversation
-          ? showSpeechRefs.originalZoomAmount
+          ? speechRefs.originalZoomAmount
           : prendyOptions.zoomLevels.default,
       });
     },
-    params: { character: "" as CharacterName, returnToZoomBeforeConversation: false },
+    params: { who: "" as CharacterName, returnToZoomBeforeConversation: undefined as boolean | undefined },
   }),
-  showMiniBubble: event({
-    run: ({ text, time }, { runMode }) => {
-      // TODO add character param
+  think: event({
+    run: ({ what: text, who, time = 100 }, { runMode }) => {
+      if (runMode !== "start") return;
 
       const { playerCharacter } = getState().global.main;
+      const character = who || playerCharacter;
       const closeCharacterMiniBubbleChainId = getCloseCharacterMiniBubbleChainId(playerCharacter);
+      setState({ miniBubbles: { [character]: { isVisible: true, text } } });
 
-      if (runMode !== "start") return;
-      setState({ miniBubbles: { [playerCharacter]: { isVisible: true, text } } });
       // 10 second timeout incase the hideMiniBubble() didn't run, like from leaving a trigger
       chainDo("cancel", closeCharacterMiniBubbleChainId);
-      runEvents([toDo("basic", "wait", { duration: time }), toDo("speech", "hideMiniBubble", {})], {
+      runEvents([II("basic", "wait", { time }), II("speech", "hideMiniBubble", {})], {
         chainId: closeCharacterMiniBubbleChainId,
       });
     },
-    params: { text: "", time: 100000 },
+    params: { what: "", who: undefined as CharacterName | undefined, time: undefined as number | undefined },
   }),
-  hideMiniBubble: event({
-    run: (_, { runMode }) => {
+  hideThink: event({
+    run: ({ who }, { runMode }) => {
+      if (runMode !== "start") return;
       const { playerCharacter } = getState().global.main;
-      setState({ miniBubbles: { [playerCharacter]: { isVisible: false } } });
+      const character = who || playerCharacter;
+      setState({ miniBubbles: { [character]: { isVisible: false } } });
     },
-    params: {},
+    params: { who: undefined as CharacterName | undefined },
   }),
-  showAlarmText: event({
-    run: async ({ text, time }, { runMode, liveId, elapsedTime }) => {
+  showTitle: event({
+    run: async ({ text, time = 1 }, { runMode, liveId, elapsedTime }) => {
       if (runMode !== "start") return;
       // NOTE alarm text in 'global' instead of project-specific 'story' ?
       setGlobalState({ alarmText: text, alarmTextIsVisible: true });
       setLiveEventState(liveId, { goalEndTime: elapsedTime + time });
       // Set it to close the next one
       chainDo("cancel", HIDE_ALARM_TEXT_CHAIN_ID);
-      runEvents([toDo("basic", "wait", { duration: time }), toDo("speech", "hideAlarmText", {})], {
+      runEvents([II("basic", "wait", { time }), II("speech", "hideTitle", {})], {
         chainId: HIDE_ALARM_TEXT_CHAIN_ID,
       });
     },
-    params: { text: "", time: 1000 },
+    params: { text: "", time: undefined as number | undefined },
   }),
-  hideAlarmText: event({
+  hideTitle: event({
     run: (_, { runMode }) => {
       if (runMode !== "start") return;
       setGlobalState({ alarmTextIsVisible: false });