+++ src/helpers/prendyHelpers/players.ts
@@ -17,12 +17,6 @@ export async function enableMovement(canMove: boolean = true, revertDelay?: numb
   }
 }
 
-// toDo move to utils/getters
-export function isHolding(pickupName: PickupName) {
-  const { heldPickups } = getState().global.main;
-  return heldPickups.includes(pickupName);
-}
-
 export function takePickup(pickup: PickupName, toHolding: boolean = true) {
   setGlobalState((state) => ({
     heldPickups: toHolding