+++ src/helpers/prendyUtils/player.ts
@@ -1,0 +1,8 @@
+import { getState } from "repond";
+import { PickupName } from "../../types";
+
+// TODO move to utils/getters
+export function isHolding(pickupName: PickupName) {
+  const { heldPickups } = getState().global.main;
+  return heldPickups.includes(pickupName);
+}